+++ src/main/ml-modules/root/runDuringDeployment/generateRemainingSearchTerms.mjs
@@ -115,15 +115,16 @@ const searchTermsConfig = {};
         return;
       }
 
-      const facetIndexReference = FACETS_CONFIG[facetName].indexReference;
+      const facetIdIndexReferences = FACETS_CONFIG[facetName].indexReferences;
+      if (facetIdIndexReferences.length > 1) {
+        return; // don't generate a search term for facets with multiple index references.
+      }
       const isIdTerm = termName.endsWith('Id');
-      const isDate = facetIndexReference.endsWith('DateLong');
-      const isDimension = facetIndexReference.endsWith('DimensionValue');
+      const isDate = facetIdIndexReferences[0].endsWith('DateLong');
+      const isDimension = facetIdIndexReferences[0].endsWith('DimensionValue');
       const isZeroOrOne = ['hasDigitalImage', 'isOnline'].includes(termName);
 
-      let facetIdIndexReferences = [facetIndexReference];
       if (isDate) {
-        facetIdIndexReferences = [facetIndexReference];
         facetIdIndexReferences.push(
           facetIdIndexReferences[0].replace('Start', 'End')
         );